Remove menu state, Add winner state
Looks like no need for pause in a card game (that was stupid) so the
pause menu was removed.
Of curse the player needs to know whether he won or lost, that's why the
winner state was added, for now it only shows you win/you lose. next
thing is to show the score.